Documentação atualizada

+++ OpenMP/src/multMatrix.h
@@ -64,11 +64,12 @@ float *multMatrix(int lin1, int col1, float *matrix1, int lin2, int col2, float 
  * 
  * @return Soma de todos os elementos da matriz
  */ 
-float reductionSum(int lin, int col, float *matrix)
+double reductionSum(int lin, int col, float *matrix)
 {
     float sum = 0;
 
     #pragma parallel for num_threads(4) reduction(+:sum)
+    // Itrando pelos elementos da matriz
     for (int i = 0; i < lin*col; i++)
     {
         sum += matrix[i];
